agentcore_agents/validation_agent/requirements.txt
strands-agents
boto3
requests
bedrock-agentcore
